extract-vsx.c: Replace "vector long" with "vector long long" throughout.
2014-03-04  Bill Schmidt  <wschmidt@linux.vnet.ibm.com>

	* gcc.dg/vmx/extract-vsx.c: Replace "vector long" with "vector
	long long" throughout.
	* gcc.dg/vmx/extract-vsx-be-order.c: Likewise.
	* gcc.dg/vmx/insert-vsx.c: Likewise.
	* gcc.dg/vmx/insert-vsx-be-order.c: Likewise.
	* gcc.dg/vmx/ld-vsx.c: Likewise.
	* gcc.dg/vmx/ld-vsx-be-order.c: Likewise.
	* gcc.dg/vmx/ldl-vsx.c: Likewise.
	* gcc.dg/vmx/ldl-vsx-be-order.c: Likewise.
	* gcc.dg/vmx/merge-vsx.c: Likewise.
	* gcc.dg/vmx/merge-vsx-be-order.c: Likewise.
	* gcc.dg/vmx/st-vsx.c: Likewise.
	* gcc.dg/vmx/st-vsx-be-order.c: Likewise.
	* gcc.dg/vmx/stl-vsx.c: Likewise.
	* gcc.dg/vmx/stl-vsx-be-order.c: Likewise.

From-SVN: r208321

+++ gcc/testsuite/gcc.dg/vmx/st-vsx-be-order.c
@@ -4,7 +4,7 @@
 
 #include "harness.h"
 
-static unsigned long svul[2] __attribute__ ((aligned (16)));
+static unsigned long long svul[2] __attribute__ ((aligned (16)));
 static double svd[2] __attribute__ ((aligned (16)));
 
 static void check_arrays ()
@@ -20,14 +20,14 @@ static void check_arrays ()
 static void test ()
 {
 #if __BYTE_ORDER__ == __ORDER_LITTLE_ENDIAN__
-  vector unsigned long vul = {1,0};
+  vector unsigned long long vul = {1,0};
   vector double vd = {1.0,0.0};
 #else
-  vector unsigned long vul = {0,1};
+  vector unsigned long long vul = {0,1};
   vector double vd = {0.0,1.0};
 #endif
 
-  vec_st (vul, 0, (vector unsigned long *)svul);
+  vec_st (vul, 0, (vector unsigned long long *)svul);
   vec_st (vd,  0, (vector double *)svd);
 
   check_arrays ();
